(WIP) Using `@vercel/swr` for client<->server states
This also implements a custom controller for the cart to help with things as I am going to make the user's cart a Map rather than an Array. Maps cannot be stored or sent over HTTP so we need some utils to use these correctly, this should (hopefully) be all done and just needs to be implemented in the real store pages. The store will not work in this commit. POC can be seen in /store/test